:truck: :: 코드 분리
event, Delegate, DataSource
