+++ Projects/Shared/UtilityModule/Sources/Extensions/Extension+Notification.Name.swift
@@ -13,4 +13,6 @@ public extension Notification.Name {
     static let refreshBookMarkGroup = Notification.Name("refreshBookMarkGroup")
     static let refreshBookMark = Notification.Name("refreshBookMark")
     static let refreshUserInfo = Notification.Name("refreshUserInfo")
+    static let filterToggle = Notification.Name("filterToggle")
+    static let refreshFilter = Notification.Name("refreshFilter")
 }
